Added profiles to testing beans, fixed property, renamed config package
