--- hermes-consumers/src/main/resources/application.yaml
@@ -1,1 +1,0 @@
-moduleName: "consumer"
